requirements.txt
discord.py==2.4.0
anthropic==0.40.0
chromadb==0.5.23
python-dotenv==1.0.1
aiohttp==3.10.11
tiktoken==0.8.0
